chore: add FakeImage to fake_image
