+++ pkg/time/fake_image.go
@@ -1,0 +1,35 @@
+// Copyright 2022 Chaos Mesh Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package time
+
+// vdsoEntryName is the name of the vDSO entry
+const vdsoEntryName = "[vdso]"
+
+// FakeImage introduce the replacement of VDSO ELF entry and customizable variables.
+// FakeImage could be constructed by LoadFakeImageFromEmbedFs(), and then used by FakeClockInjector.
+type FakeImage struct {
+	// symbolName is the name of the symbol to be replaced.
+	symbolName string
+	// content presents .text section which has been "manually relocation", the address of extern variables have been calculated manually
+	content []byte
+	// offset stores the table with variable name, and it's address in content.
+	// the key presents extern variable name, ths value is the address/offset within the content.
+	offset map[string]int
+}
+
+func NewFakeImage(symbolName string, content []byte, offset map[string]int) *FakeImage {
+	return &FakeImage{symbolName: symbolName, content: content, offset: offset}
+}
